requirements.txt
fastapi
uvicorn
python-dotenv
numpy
pydantic
pydantic-settings
pydantic[email]
ultralytics
python-multipart
pytest
httpx
asyncio
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]